Refactor: layout components and remove footer layout .
